refactor(ui): split identityManager.ts (1181 lines) into 4 responsibility-based modules identityManager.ts had 5 responsibilities (Add Form, Edit Flow, Validation, Persistence, UI Utilities) in a single 1181-line file. Split into: - identityFormValidation.ts (274 lines): field validators - identityFormUtils.ts (414 lines): shared types, UI helpers,   persistence, QuickInput infrastructure - identityAddForm.ts (334 lines): Add Identity Form logic - identityEditFlow.ts (296 lines): Edit Profile Flow logic - identityManager.ts (11 lines): barrel re-export for backward   compatibility Also fixed: removed duplicate securityLogger.logConfigChange calls from UI layer (identity layer is SSoT for audit logging).
🖥️ IDE: [Cursor](https://cursor.sh)
🔌 Extension: [Claude Code](https://claude.ai/download)

Model-Raw: claude-opus-4-6
